finish show database

+++ src/main/java/cn/edu/thssdb/parser/ImpVisitor.java
@@ -58,6 +58,7 @@ public class ImpVisitor extends SQLBaseVisitor<Object> {
         if (ctx.create_table_stmt() != null) return new QueryResult(visitCreate_table_stmt(ctx.create_table_stmt()));
         if (ctx.drop_table_stmt() != null) return new QueryResult(visitDrop_table_stmt(ctx.drop_table_stmt()));
         if (ctx.show_meta_stmt() != null) return new QueryResult(visitShow_meta_stmt(ctx.show_meta_stmt()));
+        if (ctx.show_table_stmt() != null) return new QueryResult(visitShow_table_stmt(ctx.show_table_stmt()));
         if (ctx.insert_stmt() != null) return new QueryResult(visitInsert_stmt(ctx.insert_stmt()));
         if (ctx.delete_stmt() != null) return new QueryResult(visitDelete_stmt(ctx.delete_stmt()));
         if (ctx.update_stmt() != null) return new QueryResult(visitUpdate_stmt(ctx.update_stmt()));
@@ -226,7 +227,11 @@ public class ImpVisitor extends SQLBaseVisitor<Object> {
     @Override
     public String visitShow_table_stmt(SQLParser.Show_table_stmtContext ctx) {
         try {
-            return GetCurrentDB().toString();
+            StringJoiner joiner = new StringJoiner("\n");
+            ArrayList<String> tableNames = manager.get(ctx.database_name().getText()).getTableNameList();
+            for (String tableName : tableNames)
+                joiner.add(tableName);
+            return joiner.toString();
         }
         catch (Exception e) {
             return e.getMessage();
